# server/build/swagger.json
{
	"components": {
...
		"schemas": {
...
			"MemberPartialDto": {
				"properties": {
					"lastChange": {
						"type": "string",
						"format": "date-time"
					},
					"department": {
						"$ref": "#/components/schemas/DepartmentPartialDto"
					},
					"memberStatus": {
						"$ref": "#/components/schemas/MemberStatus"
					},
					"jbtEmail": {
						"type": "string"
					},
					"mobile": {
						"type": "string"
					},
					"internalProjectId": {
						"type": "number",
						"format": "double"
					},
					"generationId": {
						"type": "number",
						"format": "double"
					},
					"lastname": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"memberId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"lastChange",
					"department",
					"memberStatus",
					"jbtEmail",
					"mobile",
					"internalProjectId",
					"generationId",
					"lastname",
					"firstname",
					"memberId"
				],
				"type": "object",
				"description": "Type of the partial member dto used for the member list"
			},
			"MemberImage": {
				"properties": {
					"mimeType": {
						"type": "string"
					},
					"base64": {
						"type": "string"
					}
				},
				"required": [
					"mimeType",
					"base64"
				],
				"type": "object",
				"description": "Type of the member image"
			},
...
			"CreateMemberResponseDto": {
				"properties": {
					"statusOverview": {
						"$ref": "#/components/schemas/StatusOverview"
					},
					"memberId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"statusOverview",
					"memberId"
				],
				"type": "object",
				"description": "Type of the response after creating a member"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "0.1.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/members": {
			"get": {
				"operationId": "GetMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MemberPartialDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of all members",
				"summary": "Get all members",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateMemberResponseDto"
								}
							}
						}
					}
				},
				"description": "Creates a new member in the database and creates accounts\r\nfor the different systems (exchange, mediawiki)",
				"summary": "Create a new member",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": [
							"1"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The member to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMemberRequestDto",
								"description": "The member to create"
							},
							"example": {
								"vorname": "Max",
								"nachname": "Mustermann",
								"name": "m.mustermann",
								"geburtsdatum": "1990-01-01",
								"password": "s3cre7-123",
								"handy": "0123456789",
								"geschlecht": 1,
								"generation": 3,
								"traineeSeit": "2020-01-01",
								"email": "m.mustermann@email.de"
							}
						}
					}
				}
			}
		},
		"/members/{id}/image": {
			"get": {
				"operationId": "GetMemberImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mimeType": {},
										"base64": {}
									},
									"required": [
										"mimeType",
										"base64"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieves the image of a member with the given `id`.\r\nReturns null with status code 204 if no image was found.",
				"summary": "Get image of member",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The id of the member to retrieve the image from",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "SaveImage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Saves the image of a member with the given `id`",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The id of the member to save the image to",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The image to save",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemberImage",
								"description": "The image to save"
							},
							"example": {
								"base64": "test",
								"mimeType": "jpg"
							}
						}
					}
				}
			}
		},
		"/members/department-members": {
			"get": {
				"operationId": "GetMembersOfDepartments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DepartmentMemberDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all members of the departments (this does not include the director of a department)",
				"summary": "Get all members of departments",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/members/directors": {
			"get": {
				"operationId": "GetDirectors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DirectorDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all current directors (if query parameter `current` is true) or all directors",
				"summary": "Get directors",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Query parameter to specify if only the current directors should be retrieved",
						"in": "query",
						"name": "current",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/members/director-positions": {
			"get": {
				"operationId": "GetDirectorPositions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DirectorPositionDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all director positions",
				"summary": "Gets all director positions",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/members/director-positions/details": {
			"get": {
				"operationId": "GetDirectorsDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DirectorDetailsDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of the current directors",
				"summary": "Get details of the current directors",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/members/{id}/director-positions/{directorPositionID}": {
			"delete": {
				"operationId": "DeleteDirectorPositions",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes the director position of the member. If the member has multiple director positions with the same id, all are deleted.",
				"summary": "Deletes the director position of the member",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": [
							"1"
						]
					}
				],
				"parameters": [
					{
						"description": "The member id",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The director position id",
						"in": "path",
						"name": "directorPositionID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "AddDirectorPosition",
				"responses": {
					"201": {
						"description": ""
					}
				},
				"description": "Adds a new director position to the member",
				"summary": "Adds a new director position",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": [
							"1"
						]
					}
				],
				"parameters": [
					{
						"description": "The id of the member to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The id of the director role",
						"in": "path",
						"name": "directorPositionID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The dates of the director position to add",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddDirectorPositionRequestDto",
								"description": "The dates of the director position to add"
							},
							"example": {
								"from": "2024-01-01",
								"until": "2024-12-12"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateDirectorPosition",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": [
							"1"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "directorPositionID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddDirectorPositionRequestDto"
							}
						}
					}
				}
			}
		},
		"/members/{id}/status": {
			"patch": {
				"operationId": "UpdateMemberStatus",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the status of an existing member\r\nUpdate can be done by members with certain permission\r\nUpdate can be done by member himself with an additional permission",
				"summary": "Update the status of a member",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": [
							"1"
						]
					}
				],
				"parameters": [
					{
						"description": "The id of the member to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The new status of the member",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"memberStatus": {
										"type": "string"
									}
								},
								"required": [
									"memberStatus"
								],
								"type": "object",
								"description": "The new status of the member"
							},
							"example": {
								"memberStatus": "aktives Mitglied"
							}
						}
					}
				}
			}
		},
		"/members/{id}/director-positions": {
			"get": {
				"operationId": "GetMemberDirectorPositions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MemberDirectorPositionsDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all directory roles of the user",
				"summary": "Get directory roles of user",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/members/departments": {
			"get": {
				"operationId": "GetDepartments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DepartmentDetailsDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all departments",
				"summary": "Get departments",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/members/departments/{id}": {
			"put": {
				"operationId": "UpdateDepartment",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the information of an existing department",
				"summary": "Update a department",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The id of the department to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateDepartmentDto"
							},
							"example": {
								"linkObjectivePresentation": "https://example.com",
								"linkOrganigram": "https://example.com"
							}
						}
					}
				}
			}
		},
		"/members/languages": {
			"get": {
				"operationId": "GetLanguages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LanguageValue"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves the language values",
				"summary": "Get languages",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/members/edv-skills": {
			"get": {
				"operationId": "GetEDVSkills",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ItSkillsValue"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves the edv skills",
				"summary": "Get edv skills",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}